rpi/scripts/screen/requirements.txt
luma.oled
psutil
pillow